Implement RawValueArray
We represent primitives as a list of objects pointing to materials and
shapes, so need to support arrays.

In the process of using this to construct the scene, I ran into an issue
due to our use of trait objects (e.g. dyn Shape) in scenes. They cannot
be cloned, so constructing primitives in a loop like this doesn't seem
to be possible (since multiple primitives could be pointing to the same
shape/material).

We have two choices:
- Start using references. This would require dealing with lifetimes.
- Switch to using enums instead of trait objects. This would cause some
  copying of data and slightly awkward implementations.