Fix the camera's co-ordinate system
I had hackishly changed how the camera converted pixels to rays to
account for the flipped y axis on screen and the right handed
co-ordinate system in OBJ files.

This commit cleans up those hacks by standardizing on a left-handed
co-ordinate system. We mirror vertices and normals in OBJ files along
the Z axis to account for the flipped co-ordinate system. This also
required a tweak in calculating triangle normals (though maybe a better
fix is to handle flipped normals correctly in shading).

The camera's job is also simplified to make this easier to understand.
It's only supposed to generate rays now, not do the sampling and
integration of irradiance. This will be useful when we add different
sampling/integration strategies.

It's still pretty confusing how the camera has to do transformations
from screen space to world space. I'll have to bite the bullet and
introduce transformation matrices to make this cleaner.