Implement lights!
I finally followed how the path tracer in pbrt works and implemented the
same approach. It's missing some pieces like MIS and computing Le when
escaping into the environment due to area lights.

The lack of area lights and MIS makes the renders worse, so should work
on that soon.