Implement light sampling using multiple importance sampling
This implements the first half of `estimate_direct`, which is to sample
sample the given light source. We don't sample the BRDFs yet, partly
because we don't have any non-delta pdf lights yet.

This also seems to occasionally fail an assert because cos_theta_i ends
up being negative, which is likely due to a missing self-intersection.