Implement BRDF sampling
Finally understood (I think) and implementing sampling of the material's
BRDF to estimate direct lighting contribution. Also fixed a bug where we
were double counting delta lights' contribution in light sampling.

The current set of lights is not very good for testing this, since two
of them are delta distributed and the infinite light is perfectly
diffuse. Will need to implement area lights for a better test.