Refactor sampling to use a shared rng provided by `render`
We now provide a Rng implementation as a type arg to render, which is
then threaded through to the various sampling functions. This allows us
to swap out the source for random numbers, which will be helpful in the
future if we want to use low discrepancy sampling etc. We now also make
use of the various sphere/disk sampling distributions that are already
provided in the library instead of rolling our own.

For now I've picked `SmallRng`, which is supposed to be fast. All of
these changes improve the render time of materials.cry by 25%.