Fix a bug in ray transformation
When rendering a mesh inside a transparent sphere, only half of the mesh
was showing up. Eventually realized that the transformed obj_ray in
sphere intersection was finding intersections that were not in the range
of the original ray. The reason was that the max_distance was not being
carried over during the transformation.