Add support for (spherical) area lights
The implementation can be improved by an alternative sampling strategy for sampling spheres. It's a little too complicated, and we would probably get better utility from disc lights, so haven't done it. Also did not implement it for triangles. That might be worth adding if we start loading meshes with emissive properties.