Add a UniformSampler
After struggling to implement a stratified sampler, I've reduced it down
to a uniform sampler, which seemed useful for debugging. The challenge
with a stratified sampler is having to permute samples. PBRT uses its
own invertible hash function, so it can generate the permutations on the
fly. We will have to generate them in advance and store them for each
dimension, which would likely be too slow to be worthwhile. Will try to
implement a different sampler.