Fix negative cos_theta
After the last fix, it became possible that w_i and normal would not be in the same hemisphere when sampling the brdf, since normals can be in arbitrary hemispheres and we were no longer relying on them to incorrectly decide whether to reflect or refract. So now we can simply do .abs on the cos_theta instead.