Rewrite path tracer from scratch
The old one was quite confusing and likely buggy. There was at least one
bug that was causing the illumination from area lights not being
estimated correctly, which caused fireflies that did increased in
brightness on increasing the number of samples.

The new one was implemented by taking `simple_integrator` and adding a
few things to it:
- MIS for the direct light arriving at an intersection
- MIS for infinite lights when the ray escapes into the scene
- MIS for an area light when the ray intersects one
- Russian roulette to terminate paths