Fix sampling issue in simple.cry
Since we now sample lights by power, a scene with infinite lights can poorly sample smaller lights if the scene is very large. This scene had a very large sphere as the base, which caused the small area light to not be sampled very often. It has been replaced with a more reasonably sized disk.