Add Image texture
The `image` crate makes this fairly easy to do. There's no mipmapping
yet, which will anyway likely require ray differentials. The initial
implementation was really slow, which turned out to be caused by the ==
comparison for BxDFs in `get_relevant_bxdfs`, which was memcmp-ing the
images in image textures. That was easily fixed by passing around
indexes, but I have also removed the `PartialEq` derives on most places
to avoid this footgun in the future.