Fix color conversion for textures
The textures were appearing washed out in all the renders, which turned
out to be because of the gamma correction we do. So we do the inverse
when converting texture values to Color. This is pretty rudimentary, but
works for now.

Also fixed the staircase scene to use the correct texture for the
stairs.